Revert "Changes to incorporate support for both IPv4 and IPv6 support"
This reverts commit cd11916dd69f4dac683a3b3903e376c0b75f18a9.
